Use gradient bitmaps for poxedex-compose images
Gradient bitmaps ensure we don't hit system-level caches while getting reproducible image resources without disk and network IO. The bitmaps are generated on the fly from the MockWebServer providing fake data.

We are using Glide's OkHttp integration in order to configure the OkHttpClient's certificates, which are used when communicating with the MockWebServer.

Test: Ran PokedexActivity on emulator, scrolled, click on item to open detail screen

Change-Id: I130d53d5d3602e4d403b4b2f18245143d948d07f