Separate domain and presentation models in Poxedex-Compose
We want to include/exclude data in the presentation layer only, like the image url. The image url depends on the mock web server's port, which changes on every server start. As such, we want to avoid caching the server url anywhere and only rely on identifiers like entity names for the requests.

We're also removing pagination logic from the database and can add it to the presentation layer if needed.

Test: Ran PokedexActivity, scrolled, opened item detail
Change-Id: Iadc41f75550ce6bfe44c22f9eb09c7789a8ebde3